refactor: remove redundant flow preview helpers
Fold createFlowPreview into gh-flow-generator and reuse GhFlowView
in the share card so flow loading/empty/error UI is not duplicated.
