+++ src/app/duckerweb/gh-flow-generator.test.ts
@@ -1,7 +1,7 @@
 import { expect, test } from "vitest";
 import fs from "node:fs";
 import { buildGhJson } from "parser/src/parser";
-import { generateFlowData } from "./gh-flow-generator";
+import { createFlowPreview, generateFlowData } from "./gh-flow-generator";
 
 test("generateFlowData creates edges with correct structure", () => {
 	const xml = fs.readFileSync("parser/sand/xmls/brep-area-Wire.xml", "utf8");
@@ -61,4 +61,17 @@ test("edge sourceHandle format is componentId.portName", () => {
 		const sourceParts = edge.sourceHandle?.split(".") ?? [];
 		expect(sourceParts.length).toBeGreaterThanOrEqual(1);
 	});
+});
+
+test("createFlowPreview creates graph data from valid Grasshopper XML", () => {
+	const xml = fs.readFileSync("parser/sand/xmls/brep-area-Wire.xml", "utf8");
+	const preview = createFlowPreview(xml);
+
+	expect(preview).not.toBeNull();
+	expect(preview?.nodes.length).toBeGreaterThan(0);
+	expect(preview?.edges.length).toBeGreaterThan(0);
+});
+
+test("createFlowPreview returns null for invalid XML", () => {
+	expect(createFlowPreview("not Grasshopper XML")).toBeNull();
 });
